refactor: 이미지 avif로 형변환, design: 페이지 selectionResult ui 개선
